requirements.txt
alembic==1.4.3
arabic-reshaper==2.1.1
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
cssselect2==0.4.1
defusedxml==0.7.1
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
flask-paginate==0.7.1
Flask-Principal==0.4.0
flask-roles==0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-User==1.0.2.2
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
future==0.18.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mimerender==0.6.0
mysql-connector-python==8.0.21
passlib==1.7.4
pdfkit==0.6.1
Pillow==8.0.0
protobuf==3.13.0
pycparser==2.20
PyJWT==1.7.1
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-editor==1.0.4
python-mimeparse==1.6.0
pytz==2020.5
reportlab==3.5.65
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.20
tinycss2==1.1.0
visitor==0.1.3
WeasyPrint==52.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
xhtml2pdf==0.2.5
